refactor theming tests
